Create HomePage and ProjectCard component
Project Card

The Project card component display one project. It recieves the `project` object
as prop. Using the information in props, it renders a project card containing project image,
title, stackList, rating and user.

HomePage

The HomePage component is connected with the redux store to dispatch the action to fetch all
project and recieve all the projects from redux store. It renders the Project Card component for each
project.