Add Project Details Page and Favorite Functionality
Project Details Page

The project details page display the information about an individual project. It display the image,
the user, rating, favorite, and description of the project. Additionality, it implements the call to action
for any project.

Favorite Functionality

Added Functionality that allows user to favorite any project. Once the user clicks favorite icon, an ajax request
is sent to the server updating the information.