Implement Favorite Component
Favorite is a container componet. On first render, it fetches the list of
all favorite projects for current user. Implemented the featchFavorites reducer and
the thunkFunction to fetch all favorite projects and dispatch to the store. The thunkFunction
uses favoriteProjects function that this implemented as a method of BackendAPI module. The favoriteProjects
function fires an ajax reques to the endpoint of favorite projects.