Add Component to Render Coversations
The component ConversationLists renders all the components and listens for new conversations.
To list all the conversations, the component fetches them from backend on the first render. Also,
it uses websockets to establish a duplex connection and subscribe to conversations channel on backend. Which
enables to update the new conversations in real time.