Change the logic to handle channel unsubscription
Previously, the channel was unsubscribed at every unmount. This resulted
in handleMessage return void. The unsubscription is now handled using
the consumer.subscriptions object. The logic checks if the
subscriptions array already contains a subscription with a
'ConversationsChannel' in case of conversation sub. In case of message
sub, it checks if sub exits with channel name and conversation id.